More refactoring as per code review
