+++ src/test/java/io/stargate/sgv2/jsonapi/api/v1/util/DataApiNamespaceCommandSender.java
@@ -1,10 +1,7 @@
 package io.stargate.sgv2.jsonapi.api.v1.util;
 
-import static io.restassured.RestAssured.given;
-
-import io.restassured.http.ContentType;
-import io.restassured.response.ValidatableResponse;
-import io.stargate.sgv2.jsonapi.api.v1.NamespaceResource;
+import io.restassured.specification.RequestSpecification;
+import io.stargate.sgv2.jsonapi.api.v1.CollectionResource;
 
 public class DataApiNamespaceCommandSender
     extends DataApiCommandSenderBase<DataApiNamespaceCommandSender> {
@@ -12,28 +9,11 @@ public class DataApiNamespaceCommandSender
     super(namespace);
   }
 
-  /**
-   * "Untyped" method for sending a POST command to the Data API: caller is responsible for
-   * formatting the JSON body correctly.
-   *
-   * @param jsonBody JSON body to POST
-   * @return Response validator for further assertions
-   */
-  public DataApiResponseValidator postRaw(String jsonBody) {
-    ValidatableResponse response =
-        given()
-            .port(getTestPort())
-            .headers(getHeaders())
-            .contentType(ContentType.JSON)
-            .body(jsonBody)
-            .when()
-            .post(NamespaceResource.BASE_PATH, namespace)
-            .then()
-            .statusCode(expectedHttpStatus.getStatusCode());
-    return new DataApiResponseValidator(response);
+  protected io.restassured.response.Response postInternal(RequestSpecification request) {
+    return request.post(CollectionResource.BASE_PATH, namespace);
   }
 
   public DataApiResponseValidator postCreateTable(String tableDefAsJSON) {
-    return postRaw("{ \"createTable\": %s }".formatted(tableDefAsJSON));
+    return postCommand("createTable", tableDefAsJSON);
   }
 }
